rework style checking workflow
This patch changes the way that we generate a directory containing the
result of a prospective merge. Previously we reset to the PR target
branch, downloaded the diff provided by GitHub for the PR, then applied
it with "git apply". It seems sometimes the diff GitHub provides doesn't
apply cleanly. We were not noticing these failures and so were running
the style checking on an unmodified copy of the PR target branch,
resulting in inaccurate results.

The flow has now been changed to perform a git merge without committing,
which is closer (should be identical) to what GitHub will perform when
hitting the merge button.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>