refactor review/checks functions
In GitHub world we're posting checks, and check() is only called by
review(), so just merge the two and name it check().

While we're at it, now we create a check at the start and update it to
completion when done. Shows up in the GH UX all nice now.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>