fix autoclosing
With the change to a GitHub App we now need to keep track of the
installation ID we are running under. In order to get a GitHub API
client instance we need to be in the corret Flask context and pass the
installation id identifying the instance we want. Otherwise we get no
client.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>